🛼 improve(minor): improve extensions
- 🐛 fix(minor): stop extensions from trying to import eslint-config
