+++ examples/critical-css/tailwind.config.js
@@ -1,4 +1,4 @@
-export default {
+module.exports = {
   content: ['./src/**/*', './public/**/*'],
   theme: {
     extend: {},
