+++ sources/@roots/bud-esbuild/package.json
@@ -62,8 +62,6 @@
     "@roots/bud-build": "workspace:sources/@roots/bud-build",
     "@roots/bud-framework": "workspace:sources/@roots/bud-framework",
     "@skypack/package-check": "0.2.2",
-    "@types/fs-extra": "9.0.13",
-    "@types/lodash": "4.14.182",
     "@types/node": "16.11.36",
     "webpack": "5.72.1"
   },
