+++ sources/@roots/bud-extensions/src/service.ts
@@ -131,12 +131,12 @@ export default class Extensions
       .filter(signifier => !signifier.startsWith('@types'))
       .filter(
         signifier =>
-          signifier.startsWith('@roots') || signifier.includes('bud-'),
+          signifier.startsWith('@roots/bud-') ||
+          signifier.startsWith('@roots/sage'),
       )
       .filter(
         signifier =>
           ![
-            '@roots/bud',
             '@roots/bud-api',
             '@roots/bud-build',
             '@roots/bud-cache',
@@ -147,7 +147,6 @@ export default class Extensions
             '@roots/bud-framework',
             '@roots/bud-hooks',
             '@roots/bud-server',
-            '@roots/container',
           ].includes(signifier),
       )
       .filter(signifier => !this.has(signifier))
