+++ sources/@roots/bud-server/package.json
@@ -53,10 +53,7 @@
   "devDependencies": {
     "@skypack/package-check": "0.2.2",
     "@types/cookie-parser": "1.4.3",
-    "@types/cors": "2.8.12",
     "@types/express": "4.17.13",
-    "@types/lodash": "4.14.182",
-    "@types/lodash-es": "^4",
     "@types/node": "16.11.36",
     "@types/signale": "1.4.4",
     "@types/webpack-dev-middleware": "5.0.2",
@@ -69,16 +66,12 @@
     "ansi-html-community": "0.0.8",
     "chokidar": "3.5.3",
     "cookie-parser": "1.4.6",
-    "cors": "2.8.5",
     "express": "4.18.1",
     "globby": "13.1.1",
     "helpful-decorators": "2.1.0",
-    "html-entities": "^2.1.0",
     "http-proxy-middleware": "2.0.6",
     "lodash-es": "4.17.21",
-    "querystring": "^0.2.0",
     "signale": "1.4.0",
-    "strip-ansi": "^7.0.0",
     "tslib": "2.4.0",
     "webpack-dev-middleware": "5.3.3"
   }
