+++ sources/@roots/bud-stylelint/src/extension.ts
@@ -1,0 +1,51 @@
+import {Extension} from '@roots/bud-framework/extension'
+import {
+  bind,
+  expose,
+  label,
+  options,
+  plugin,
+} from '@roots/bud-framework/extension/decorators'
+import StylelintWebpackPlugin, {Options} from 'stylelint-webpack-plugin'
+
+/**
+ * Bud stylelint extension
+ *
+ * @public
+ * @decorator `@label`
+ * @decorator `@plugin`
+ * @decorator `@options`
+ * @decorator `@expose`
+ */
+@label('@roots/bud-stylelint')
+@plugin(StylelintWebpackPlugin)
+@options({context: app => app.path('@src')})
+@expose('stylelint')
+export default class BudStylelintWebpackPlugin extends Extension<
+  Options,
+  StylelintWebpackPlugin
+> {
+  /**
+   * Fail build on stylelint error
+   *
+   * @public
+   * @decorator `@bind`
+   */
+  @bind
+  public failOnError(fail: boolean = true): this {
+    this.options.failOnError = fail
+    return this
+  }
+
+  /**
+   * Fail build on stylelint warning
+   *
+   * @public
+   * @decorator `@bind`
+   */
+  @bind
+  public failOnWarning(fail: boolean = true): this {
+    this.options.failOnWarning = fail
+    return this
+  }
+}
