+++ sources/@roots/bud-stylelint/src/index.ts
@@ -10,41 +10,7 @@
  * @packageDocumentation
  */
 
-import {Extension} from '@roots/bud-framework'
-import {
-  bind,
-  expose,
-  label,
-  options,
-  plugin,
-} from '@roots/bud-framework/extension/decorators'
-import StylelintWebpackPlugin, {Options} from 'stylelint-webpack-plugin'
+import './env.js'
 
-declare module '@roots/bud-framework' {
-  interface Bud {
-    stylelint: BudStylelintWebpackPlugin
-  }
-  interface Modules {
-    '@roots/bud-stylelint': BudStylelintWebpackPlugin
-  }
-}
-
-@label('stylelint-webpack-plugin')
-@plugin(StylelintWebpackPlugin)
-@options({context: app => app.path('@src')})
-@expose('stylelint')
-export default class BudStylelintWebpackPlugin extends Extension<
-  Options,
-  StylelintWebpackPlugin
-> {
-  @bind
-  public failOnError(fail: boolean = true): this {
-    this.options.failOnError = fail
-    return this
-  }
-  @bind
-  public failOnWarning(fail: boolean = true): this {
-    this.options.failOnWarning = fail
-    return this
-  }
-}
+import BudStylelint from './extension.js'
+export default BudStylelint
